+++ test/python/test_featnorm.py
@@ -1,4 +1,3 @@
-
 from typing import Tuple
 import logging
 import torch